fix: clear test-hygiene gate zero-tolerance violations
- ollama-provider.test.ts: assert format equals the schema object
  instead of toBeDefined()
- intent-parser.test.ts, schemas.test.ts: replace dead id.toBeTruthy()
  checks with a format assertion matching the nanoid id shape
- prompt-builder.test.ts: assert version format instead of toBeTruthy()
- schemas.test.ts: split ModelIntentPacket tests into their own file
  to clear the per-file test cap
